changed to SUNDIALS v5.1.0
Changed the src and include files to SUNDIALS v5.1.0. Added a NEWS file

+++ src/sundials/ida/ida_ls_impl.h
@@ -2,19 +2,15 @@
  * Programmer(s): Daniel R. Reynolds @ SMU
  *                Alan C. Hindmarsh and Radu Serban @ LLNL
  *-----------------------------------------------------------------
- * LLNS/SMU Copyright Start
- * Copyright (c) 2018, Southern Methodist University and
- * Lawrence Livermore National Security
- *
- * This work was performed under the auspices of the U.S. Department
- * of Energy by Southern Methodist University and Lawrence Livermore
- * National Laboratory under Contract DE-AC52-07NA27344.
- * Produced at Southern Methodist University and the Lawrence
- * Livermore National Laboratory.
- *
+ * SUNDIALS Copyright Start
+ * Copyright (c) 2002-2020, Lawrence Livermore National Security
+ * and Southern Methodist University.
  * All rights reserved.
- * For details, see the LICENSE file.
- * LLNS/SMU Copyright End
+ *
+ * See the top-level LICENSE and NOTICE files for details.
+ *
+ * SPDX-License-Identifier: BSD-3-Clause
+ * SUNDIALS Copyright End
  *-----------------------------------------------------------------
  * Implementation header file for IDA's linear solver interface.
  *-----------------------------------------------------------------*/
@@ -73,7 +69,7 @@ typedef struct IDALsMemRec {
   long int ncfl0;     /* ncfl0 = saved ncfl (for performance monitor) */
   long int nwarn;     /* nwarn = no. of warnings (for perf. monitor)  */
 
-  long int last_flag; /* last error return flag                       */
+  int last_flag;      /* last error return flag                       */
 
   /* Preconditioner computation
      (a) user-provided:
